Cleared up out usage of `.gitkeep` files
It is not clear when or why these were added, but there should not be
any need to keep these directories in version control.

I suspect the interns had an issue from a clean working directory and
this is how they attempted to resolve it. I will reset my working
directory as best as I can and see if there is a permissions problem on
a clean startup, then resolve it within the start script.