Removed usage of `~` path
Relative paths are safer, and best practice. Not sure how this was even
working.
I assumed this would only work if the container viewed the user
directory as the app directory but this isn't the case... Not sure how
it was resolving, but possibly relying on default behavior during a
failed path lookup.