Cleaned up all styling and a few CSS issues
There was a gradient on the page that was being creating incorrectly. I
moved that to the main app as a classAttr on the `body` element.

There was also an attempt at some left and right margin for improved
readability near the edges of the screen. I couldn't fully follow the
process so I might have not completely removed the bad code, but it was
causing unnecessary scrollbars in some cases and I fixed it with an
application wide style to the main Nuxt `div#__nuxt` element.